fix(gateway): keep cron result schemas within module budget
